refactor(AppStructure): rename and adjust to common style
1. Move Angular related class to `'./angular'`.
2. Add newlines/spaces to adjust to common style.
