+++ src/angular/ng2-walker.ts
@@ -4,69 +4,89 @@ import * as compiler from '@angular/compiler';
 import { TemplateAst, ElementAst, AttrAst, TextAst, BoundTextAst, BoundElementPropertyAst, BoundEventAst } from '@angular/compiler';
 import * as e from '@angular/compiler/src/expression_parser/ast';
 
-import {parseTemplate} from './templateParser';
+import {parseTemplate} from './template-parser';
 
 const ExpTypes = {
   Binary(ast: any) {
     return ast.constructor.name === 'Binary';
   },
+
   Quote(ast: any) {
     return ast.constructor.name === 'Quote';
   },
+
   EmptyExpr(ast: any) {
     return ast.constructor.name === 'EmptyExpr';
   },
+
   ImplicitReceiver(ast: any) {
     return ast.constructor.name === 'ImplicitReceiver';
   },
+
   Chain(ast: any) {
     return ast.constructor.name === 'Chain';
   },
+
   Conditional(ast: any) {
     return ast.constructor.name === 'Conditional';
   },
+
   PropertyRead(ast: any) {
     return ast.constructor.name === 'PropertyRead';
   },
+
   PropertyWrite(ast: any) {
     return ast.constructor.name === 'PropertyWrite';
   },
+
   SafePropertyRead(ast: any) {
     return ast.constructor.name === 'SafePropertyRead';
   },
+
   KeyedRead(ast: any) {
     return ast.constructor.name === 'KeyedRead';
   },
+
   KeyedWrite(ast: any) {
     return ast.constructor.name === 'KeyedWrite';
   },
+
   BindingPipe(ast: any) {
     return ast.constructor.name === 'BindingPipe';
   },
+
   LiteralPrimitive(ast: any) {
     return ast.constructor.name === 'LiteralPrimitive';
   },
+
   LiteralArray(ast: any) {
     return ast.constructor.name === 'LiteralArray';
   },
+
   LiteralMap(ast: any) {
     return ast.constructor.name === 'LiteralMap';
   },
+
   Interpolation(ast: any) {
     return ast.constructor.name === 'Interpolation';
   },
+
   PrefixNot(ast: any) {
     return ast.constructor.name === 'PrefixNot';
   },
+
   MethodCall(ast: any) {
     return ast.constructor.name === 'MethodCall';
   },
+
   SafeMethodCall(ast: any) {
     return ast.constructor.name === 'SafeMethodCall';
   },
+
   FunctionCall(ast: any) {
     return ast.constructor.name === 'FunctionCall';
   },
+
   ASTWithSource(ast: any) {
     return ast.constructor.name === 'ASTWithSource';
   }
@@ -93,81 +113,102 @@ export class RecursiveAngularExpressionVisitor extends Lint.RuleWalker implement
   constructor(sourceFile: ts.SourceFile, options: Lint.IOptions, protected context: ts.ClassDeclaration, protected basePosition: number) {
     super(sourceFile, options);
   }
+
   visit(ast: e.AST, context: any) {
     ast.visit(this);
     return null;
   }
+
   visitBinary(ast: e.Binary, context: any): any {
     ast.left.visit(this);
     ast.right.visit(this);
     return null;
   }
+
   visitChain(ast: e.Chain, context: any): any { return this.visitAll(ast.expressions, context); }
+
   visitConditional(ast: e.Conditional, context: any): any {
     ast.condition.visit(this);
     ast.trueExp.visit(this);
     ast.falseExp.visit(this);
     return null;
   }
+
   visitPipe(ast: e.BindingPipe, context: any): any {
     ast.exp.visit(this);
     this.visitAll(ast.args, context);
     return null;
   }
+
   visitFunctionCall(ast: e.FunctionCall, context: any): any {
     ast.target.visit(this);
     this.visitAll(ast.args, context);
     return null;
   }
+
   visitImplicitReceiver(ast: e.ImplicitReceiver, context: any): any { return null; }
+
   visitInterpolation(ast: e.Interpolation, context: any): any {
     return this.visitAll(ast.expressions, context);
   }
+
   visitKeyedRead(ast: e.KeyedRead, context: any): any {
     ast.obj.visit(this);
     ast.key.visit(this);
     return null;
   }
+
   visitKeyedWrite(ast: e.KeyedWrite, context: any): any {
     ast.obj.visit(this);
     ast.key.visit(this);
     ast.value.visit(this);
     return null;
   }
+
   visitLiteralArray(ast: e.LiteralArray, context: any): any {
     return this.visitAll(ast.expressions, context);
   }
+
   visitLiteralMap(ast: e.LiteralMap, context: any): any { return this.visitAll(ast.values, context); }
+
   visitLiteralPrimitive(ast: e.LiteralPrimitive, context: any): any { return null; }
+
   visitMethodCall(ast: e.MethodCall, context: any): any {
     ast.receiver.visit(this);
     return this.visitAll(ast.args, context);
   }
+
   visitPrefixNot(ast: e.PrefixNot, context: any): any {
     ast.expression.visit(this);
     return null;
   }
+
   visitPropertyRead(ast: e.PropertyRead, context: any): any {
     ast.receiver.visit(this);
     return null;
   }
+
   visitPropertyWrite(ast: e.PropertyWrite, context: any): any {
     ast.receiver.visit(this);
     ast.value.visit(this);
     return null;
   }
+
   visitSafePropertyRead(ast: e.SafePropertyRead, context: any): any {
     ast.receiver.visit(this);
     return null;
   }
+
   visitSafeMethodCall(ast: e.SafeMethodCall, context: any): any {
     ast.receiver.visit(this);
     return this.visitAll(ast.args, context);
   }
+
   visitAll(asts: e.AST[], context: any): any {
     asts.forEach(ast => ast.visit(this, context));
     return null;
   }
+
   visitQuote(ast: e.Quote, context: any): any { return null; }
 }
