+++ src/angular/template-parser.ts
