+++ src/directiveClassSuffixRule.ts
@@ -1,36 +1,33 @@
 import * as Lint from 'tslint/lib/lint';
 import * as ts from 'typescript';
 import {sprintf} from 'sprintf-js';
-import SyntaxKind = require('./util/syntaxKind');
-import {Ng2Walker} from "./util/ng2Walker";
+import SyntaxKind = require('./util/syntax-kind');
+import {Ng2Walker} from './angular/ng2-walker';
 
 export class Rule extends Lint.Rules.AbstractRule {
+  public apply(sourceFile:ts.SourceFile):Lint.RuleFailure[] {
+    return this.applyWithWalker(
+      new ClassMetadataWalker(sourceFile,
+        this.getOptions()));
+  }
 
-    public apply(sourceFile:ts.SourceFile):Lint.RuleFailure[] {
-        return this.applyWithWalker(
-            new ClassMetadataWalker(sourceFile,
-                this.getOptions()));
-    }
+  static FAILURE:string = "The name of the class %s should end with the suffix Directive ($$02-03$$)";
 
-    static FAILURE:string = "The name of the class %s should end with the suffix Directive ($$02-03$$)";
-
-    static validate(className:string):boolean {
-        return /.*Directive/.test(className);
-    }
+  static validate(className:string):boolean {
+    return /.*Directive/.test(className);
+  }
 }
 
 export class ClassMetadataWalker extends Ng2Walker {
-
-    visitNg2Directive(controller:ts.ClassDeclaration, decorator:ts.Decorator) {
-        let name = controller.name;
-        let className:string = name.text;
-        if (!Rule.validate(className)) {
-            this.addFailure(
-                this.createFailure(
-                    name.getStart(),
-                    name.getWidth(),
-                    sprintf.apply(this, [Rule.FAILURE, className])));
-        }
+  visitNg2Directive(controller:ts.ClassDeclaration, decorator:ts.Decorator) {
+    let name = controller.name;
+    let className:string = name.text;
+    if (!Rule.validate(className)) {
+      this.addFailure(
+        this.createFailure(
+          name.getStart(),
+          name.getWidth(),
+          sprintf.apply(this, [Rule.FAILURE, className])));
     }
-
+  }
 }
