+++ src/useOutputPropertyDecoratorRule.ts
@@ -11,3 +11,4 @@ export class Rule extends UsePropertyDecorator {
     }, ruleName, value, disabledIntervals);
   }
 }
+
